norm: Resolve norm errors that exceeding line numbers and too many functions in a file
